pyproject.toml
[project]
name = "ipc"
version = "0.1.0"
description = "Asynchronous implementation of the Majordomo broker"
authors = [
    {name = "Void-ux", email = "77840397+Void-ux@users.noreply.github.com"},
]
dependencies = [
    "pyzmq>=26.4.0",
    "msgpack>=1.1.0",
]
requires-python = "<4.0,>=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
broker = []
client = []
worker = [
    "discord-py>=2.5.2",
]
all = ["ipc[broker,worker,client]"]


[tool.pdm]
distribution = false

[tool.pdm.build]
includes = []

[tool.pdm.dev-dependencies]
dev = []

